requirements.txt
flask[async]==2.3.3 
openai==1.25.0 
httpx==0.24.1
faiss-cpu==1.8.0 
pypdf==4.2.0 
python-docx==1.1.0 
pandas==2.2.2 
numpy==1.26.4 
python-dotenv==1.0.0
aiohttp==3.9.5
docx2txt==0.8
packaging==24.0